requirements.lock.txt
# Reproducible dependency lock for CI/docker matrix harness.
# Refresh with: python scripts/refresh_lockfile.py
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
coverage==7.13.4
googleapis-common-protos==1.72.0
grpcio==1.78.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
mypy==1.19.1
mypy_extensions==1.1.0
opentelemetry-api==1.39.1
opentelemetry-distro==0.60b1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging==26.0
pluggy==1.6.0
protobuf==6.33.5
Pygments==2.19.2
pytest==9.0.2
pytest-benchmark==5.2.3
pytest-cov==7.0.0
requests==2.32.5
rich==14.3.2
ruff==0.15.1
sentry-sdk==2.52.0
structlog==25.5.0
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==1.17.3
zipp==3.23.0
